Refactored database services for code maintainability
