--- FlashcardsAPI/Services/IFlashcardsAppDbService.cs
@@ -1,7 +1,0 @@
-﻿namespace FlashcardsAPI.Services
-{
-	// To be removed in the future
-	public interface IFlashcardsAppDbService : ICommentService, IFlashcardService, IFlashcardCollectionService, IReactionService
-	{   
-    }
-}
